// types/mail-forms.ts
export interface IContactInputs {
  name: string;
  email: string;
  message: string;
}

export interface IFeedbackInputs {
  name: string;
  email: string;
  feedback: string;
}

export interface IQuotationInputs {
  name: string;
  email: string;
  phone: string;
  service: string;
  description: string;
}